refactor: Payment 대비 폴더 구조 수정
#52
