+++ backend/order-service/src/main/java/com/bidket/order/domain/order/model/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.bidket.order.domain.model;
+package com.bidket.order.domain.order.model;
 
 public enum OrderStatus {
     PAYMENT,    // 결제 대기
