+++ backend/order-service/src/main/java/com/bidket/order/domain/order/repository/OrderRepository.java
@@ -1,0 +1,8 @@
+package com.bidket.order.domain.order.repository;
+
+import com.bidket.order.domain.order.model.Order;
+
+public interface OrderRepository {
+
+    Order save(Order order);
+}
