--- backend/order-service/src/main/java/com/bidket/order/domain/repository/OrderRepository.java
@@ -1,8 +1,0 @@
-package com.bidket.order.domain.repository;
-
-import com.bidket.order.domain.model.Order;
-
-public interface OrderRepository {
-
-    Order save(Order order);
-}
