+++ backend/order-service/src/main/java/com/bidket/order/presentation/order/dto/request/OrderCreateRequest.java
@@ -1,4 +1,4 @@
-package com.bidket.order.presentation.dto.request;
+package com.bidket.order.presentation.order.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Min;
